[fix] AI stuff

+++ include/AI.h
@@ -190,7 +190,7 @@ void AI_readJSON( void *ai, cJSON *jai);
 void AI_writeJSON(void *ai, cJSON *jai);
 
 s8 AI_filename(i32 ai_id);
-s8 AI_ID_isvalid(i32 ai_id);
+i32 AI_ID_isvalid(i32 ai_id);
 
 /* --- AI_State --- */
 void AI_State_Pop( struct AI_State *ais, tnecs_world *world);
